Split buildDefault into separate file in @codotype/util
